+++ mux/src/client.rs
@@ -34,7 +34,7 @@ impl ClientId {
             pid: unsafe { libc::getpid() as u32 },
             epoch: *EPOCH,
             id,
-            ssh_auth_sock: std::env::var("SSH_AUTH_SOCK").ok(),
+            ssh_auth_sock: crate::AgentProxy::default_ssh_auth_sock(),
         }
     }
 }