+++ mux/src/ssh_agent.rs
@@ -84,7 +84,7 @@ impl AgentProxy {
         let pid = unsafe { libc::getpid() };
         let sock_path = config::RUNTIME_DIR.join(format!("agent.{pid}"));
 
-        if let Ok(inherited) = std::env::var("SSH_AUTH_SOCK") {
+        if let Some(inherited) = Self::default_ssh_auth_sock() {
             if let Err(err) = update_symlink(&inherited, &sock_path) {
                 log::error!("failed to set {sock_path:?} to initial inherited SSH_AUTH_SOCK value of {inherited:?}: {err:#}");
             }
@@ -98,6 +98,13 @@ impl AgentProxy {
             sock_path,
             current_target: RwLock::new(None),
             sender,
+        }
+    }
+
+    pub fn default_ssh_auth_sock() -> Option<String> {
+        match &config::configuration().default_ssh_auth_sock {
+            Some(value) => Some(value.to_string()),
+            None => std::env::var("SSH_AUTH_SOCK").ok(),
         }
     }